Test: Consolidate site test utils into TestSiteSqlUtils
Move insertTestAccountAndSiteIntoDb from SiteTestUtils into
TestSiteSqlUtils to have a single site test utility class.
Delete SiteTestUtils.kt as it is now empty.
